api/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.0
feedparser>=6.0
trafilatura>=1.12
ebooklib>=0.18
Pillow>=11.0
google-api-python-client>=2.160
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
pyyaml>=6.0
python-multipart>=0.0.18
PyJWT>=2.8
cryptography>=44.0
